+++ lib/plugin/common/converter/parser.ts
@@ -1,0 +1,4 @@
+/**
+ * @hidden
+ */
+export type Parser<T> = (text: string) => T | null;
