--- lib/plugin/common/reader/parser.ts
@@ -1,4 +1,0 @@
-/**
- * @hidden
- */
-export type Parser<In, Ex> = (text: In) => Ex | null;
