+++ lib/plugin/input-bindings/color/view/h-palette.ts
@@ -1,9 +1,9 @@
-import {colorToFunctionalRgbString} from '../../../common/converter/color';
 import {Color} from '../../../common/model/color';
 import {Value} from '../../../common/model/value';
 import {mapRange} from '../../../common/number-util';
 import {ClassName} from '../../../common/view/class-name';
 import {ValueView} from '../../../common/view/value';
+import {colorToFunctionalRgbString} from '../converter/color-string';
 
 const className = ClassName('hpl');
