+++ lib/plugin/input-bindings/number/controller/number-text-test.ts
@@ -2,9 +2,11 @@ import {assert} from 'chai';
 import {describe, it} from 'mocha';
 
 import {TestUtil} from '../../../../misc/test-util';
-import {NumberFormatter} from '../../../common/converter/number';
+import {
+	NumberFormatter,
+	StringNumberParser,
+} from '../../../common/converter/number';
 import {Value} from '../../../common/model/value';
-import {StringNumberParser} from '../../../common/reader/string-number';
 import {NumberTextController} from './number-text';
 
 describe(NumberTextController.name, () => {
