+++ lib/plugin/input-bindings/point-2d/plugin.ts
@@ -7,10 +7,12 @@ import {
 import {Constraint} from '../../common/constraint/constraint';
 import {RangeConstraint} from '../../common/constraint/range';
 import {StepConstraint} from '../../common/constraint/step';
-import {NumberFormatter} from '../../common/converter/number';
+import {
+	NumberFormatter,
+	StringNumberParser,
+} from '../../common/converter/number';
 import {Value} from '../../common/model/value';
 import {PaneError} from '../../common/pane-error';
-import {StringNumberParser} from '../../common/reader/string-number';
 import {InputBindingPlugin} from '../../input-binding';
 import {getBaseStep, getSuitableDecimalDigits} from '../../util';
 import {Point2dConstraint} from './constraint/point-2d';
